refactor: split terminal and prompt preferences (#117)
Closes #117
